# [Vulkan] Optimize GRU operator with pre-packing