--- a/aten/src/ATen/native/vulkan/ops/Register.cpp
+++ b/aten/src/ATen/native/vulkan/ops/Register.cpp
@@ -2,6 +2,7 @@
 
 #include <ATen/native/vulkan/ops/Common.h>
 #include <ATen/native/vulkan/ops/Convolution.h>
+#include <ATen/native/vulkan/ops/Gru.h>
 #include <ATen/native/vulkan/ops/TransposeConvolution2d.h>
 #include <ATen/native/vulkan/ops/Mm.h>
 #include <torch/custom_class.h>
@@ -28,9 +29,9 @@ TORCH_LIBRARY(vulkan, m) {
                 std::move(std::get<2>(state)),
                 std::move(std::get<3>(state)),
                 std::move(std::get<4>(state)),
-                std::move(std::get<5>(state)),
-                std::move(std::get<6>(state)),
-                std::move(std::get<7>(state)));
+                std::get<5>(state),
+                std::get<6>(state),
+                std::get<7>(state));
           });
   m.class_<TransposeConv2dOpContext>("TransposeConv2dOpContext")
       .def_pickle(
@@ -47,9 +48,9 @@ TORCH_LIBRARY(vulkan, m) {
                 std::move(std::get<3>(state)),
                 std::move(std::get<4>(state)),
                 std::move(std::get<5>(state)),
-                std::move(std::get<6>(state)),
-                std::move(std::get<7>(state)),
-                std::move(std::get<8>(state)));
+                std::get<6>(state),
+                std::get<7>(state),
+                std::get<8>(state));
           });
   m.class_<LinearOpContext>("LinearOpContext")
       .def_pickle(
@@ -61,6 +62,23 @@ TORCH_LIBRARY(vulkan, m) {
           [](LinearOpContext::State state) {
             return linear_prepack(
                 std::move(std::get<0>(state)), std::move(std::get<1>(state)));
+          });
+  m.class_<GruOpContext>("GruOpContext")
+      .def_pickle(
+          // __getstate__
+          [](const c10::intrusive_ptr<GruOpContext>& context) {
+            return context->unpack();
+          },
+          // __setstate__
+          [](GruOpContext::State state) {
+            return gru_prepack(
+                std::move(std::get<0>(state)),
+                std::get<1>(state),
+                std::get<2>(state),
+                std::get<3>(state),
+                std::get<4>(state),
+                std::get<5>(state),
+                std::get<6>(state));
           });
 }
 
@@ -87,18 +105,33 @@ TORCH_LIBRARY(vulkan_prepack, m) {
   m.def(TORCH_SELECTIVE_SCHEMA(
       "vulkan_prepack::linear_run(Tensor X, "
       "__torch__.torch.classes.vulkan.LinearOpContext BW_prepack) -> Tensor Y"));
+  m.def(TORCH_SELECTIVE_SCHEMA(
+      "vulkan_prepack::gru_prepack(Tensor[] params_cpu, "
+      "bool has_biases, "
+      "int num_layers, "
+      "float dropout, "
+      "bool train, "
+      "bool bidirectional, "
+      "bool batch_first) "
+      "-> __torch__.torch.classes.vulkan.GruOpContext"));
+  m.def(TORCH_SELECTIVE_SCHEMA(
+      "vulkan_prepack::gru_run(Tensor input_vk, "
+      "Tensor hx_vk, "
+      "__torch__.torch.classes.vulkan.GruOpContext G_prepack) -> (Tensor next_input, Tensor hidden_layer)"));
 }
 
 TORCH_LIBRARY_IMPL(vulkan_prepack, CPU, m) {
   m.impl(TORCH_SELECTIVE_NAME("vulkan_prepack::conv2d_clamp_prepack"), TORCH_FN(conv2d_clamp_prepack));
   m.impl(TORCH_SELECTIVE_NAME("vulkan_prepack::conv2d_transpose_clamp_prepack"), TORCH_FN(conv2d_transpose_clamp_prepack));
   m.impl(TORCH_SELECTIVE_NAME("vulkan_prepack::linear_prepack"), TORCH_FN(linear_prepack));
+  m.impl(TORCH_SELECTIVE_NAME("vulkan_prepack::gru_prepack"), TORCH_FN(gru_prepack));
 }
 
 TORCH_LIBRARY_IMPL(vulkan_prepack, Vulkan, m) {
   m.impl(TORCH_SELECTIVE_NAME("vulkan_prepack::conv2d_clamp_run"), TORCH_FN(conv2d_clamp_run));
   m.impl(TORCH_SELECTIVE_NAME("vulkan_prepack::conv2d_transpose_clamp_run"), TORCH_FN(conv2d_transpose_clamp_run));
   m.impl(TORCH_SELECTIVE_NAME("vulkan_prepack::linear_run"), TORCH_FN(linear_run));
+  m.impl(TORCH_SELECTIVE_NAME("vulkan_prepack::gru_run"), TORCH_FN(gru_run));
 }
 
 Tensor convolution(